Handle HTTP connections made to a websocketport.
This patch adds support for a very minimal subset of the HTTP protocol.
We are mostly ignoring everything the clients sends our way unless it is
an HTTP protocol upgrade request for a WebSocket connection.

If the client tries to make an ordinary HTTP request, we simply shut down
the connection and responds with a 400 Bad Request HTTP status reply.

Since our HTTP protocol parser allows us to add partial strings, we use
the Ratbox linebuffers in RAW and PARTIAL mode, such that we may or may
not receive complete lines. This is also useful for us for handling
WebSocket frames later on, due to the fact that the WebSocket framing
protocol is a binary protocol and therefore we can't wait for complete
lines, since the concept of lines does not exist in that protocol.